Practice Model
This notebook, titled "practice_model_urban8k.ipynb," serves as a practical guide for building and training a machine-learning model from scratch on the UrbanSound8K dataset. The notebook covers data loading, preprocessing, model construction, training, and evaluation, providing a hands-on experience for individuals interested in audio classification tasks. 

The primary purpose of this notebook is to:
1. Demonstrate the process of creating a machine-learning model for audio classification.
2. Provide practical insights into data preprocessing and feature extraction from the UrbanSound8K dataset.
3. Offer guidance on model training and evaluation